Add riak_core as application dep to riak_api.app
When generating a release with erlang r14b0* riak fails to start.
This is due to the start order in the generated riak.script.
According to systools man page -
"The applications are sorted according to the dependencies between
the applications. Where there are no dependencies, the order in the
.rel file is kept." There may be a better way to control the
order in the rel file, but this PR fixes the issue by
declaring riak_api's dependency on riak_core.