Use folsom's `gauge` type to store the function needed to pbcconnects
PBC connects metric is the number of `active` children in
`supervisr:count_children(riak_api_pb_sup)`. However, new stats query
work expects all metrics to be found in folsom's `metrics` table,
to that end this commit means you can name / store a stat in folsom
but have the production of the value be a function.